+++ source/app/rendering/primitives/arrow.cpp
@@ -279,7 +279,7 @@ void Arrow::generateVertexData(std::vector<float>& vertices, std::vector<float>&
 
     index = 0;
 
-    for(unsigned long slice = 0U; slice < _slices; slice++)
+    for(unsigned int slice = 0u; slice < _slices; slice++)
     {
         auto baseIndex = slice * 4;
